Fix a few errors with v142 compiler
I can't fix "GetVersionEx is deprecated" because IsWindowsVistaOrGreater
doesn't exist in v141_xp (only in v141).

I didn't manage to fix the "unresolved external symbol _memset" one. I guess
we will have to either link the C runtime statically, link to MSVCRT.dll,
or implement it in assembly (where the C compiler can't tell us "don't define
this function, I already defined it!").

+++ thcrap_update/src/http_curl.cpp
@@ -27,7 +27,7 @@ size_t CurlHandle::writeCallbackStatic(char *ptr, size_t size, size_t nmemb, voi
 int CurlHandle::progressCallbackStatic(void *userdata, curl_off_t dltotal, curl_off_t dlnow, curl_off_t /* ultotal */, curl_off_t /* ulnow */)
 {
     auto callback = *static_cast<std::function<bool(size_t, size_t)>*>(userdata);
-    if (callback(dlnow, dltotal)) {
+    if (callback((size_t)dlnow, (size_t)dltotal)) {
         return 0;
     }
     else {
